Extract temporal parsing utils to shared module
Consolidates duplicate `parseDateLikeString` and date coercion logic from `PipelineStore.ts` and `fromFlintChart.ts` into a new shared `temporalStrings.ts` module. Also refactors `coerceTemporalStringRows` to return a result object with a `failed` flag instead of mutating a warnings array, separating the coercion concern from warning generation.
